extra/motor/requirements.txt
absl-py==0.11.0
appnope==0.1.2
astor==0.8.1
astunparse==1.6.3
backcall==0.2.0
braceexpand==0.1.6
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
clang==5.0
cycler==0.10.0
deap==1.3.1
decorator==4.4.2
efficientnet==1.0.0
flatbuffers==1.12
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.41.1
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
networkx==2.5
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.5
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
ply==3.11
prompt-toolkit==3.0.10
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.0
Pygments==2.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.18.0rc1
scipy==1.4.1
six==1.15.0
str2bool==1.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
tifffile==2020.9.3
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
